Getting member requesting not successful
I had tried several ways to get requested user to the event but didn't work. I commented out some functions. Let me know if you can do it cuz i gave up for now. And also getting members of the event didn't work but somehow manage to be working but the code is so lame.